Moved most of lir.io into io.lir, removed old JSON io
