+++ tla-io/src/main/scala/at/forsyte/apalache/io/annotations/PrettyWriterWithAnnotationsFactory.scala
@@ -1,7 +1,7 @@
 package at.forsyte.apalache.io.annotations
 
 import at.forsyte.apalache.io.annotations.store.AnnotationStore
-import at.forsyte.apalache.tla.lir.io.{TlaWriter, TlaWriterFactory}
+import at.forsyte.apalache.io.lir.{TlaWriter, TlaWriterFactory}
 import com.google.inject.{Inject, Singleton}
 
 import java.io.PrintWriter
