+++ tla-io/src/main/scala/at/forsyte/apalache/io/json/impl/TlaToUJson.scala
@@ -1,7 +1,7 @@
 package at.forsyte.apalache.io.json.impl
 
 import at.forsyte.apalache.io.json.TlaToJson
-import at.forsyte.apalache.tla.lir.io.TypeTagPrinter
+import at.forsyte.apalache.io.lir.TypeTagPrinter
 import at.forsyte.apalache.tla.lir.storage.SourceLocator
 
 /**
