+++ tla-io/src/main/scala/at/forsyte/apalache/io/json/impl/UJsonToTla.scala
@@ -2,7 +2,7 @@ package at.forsyte.apalache.io.json.impl
 
 import at.forsyte.apalache.io.json.JsonToTla
 import at.forsyte.apalache.tla.imp.src.SourceStore
-import at.forsyte.apalache.tla.lir.io.TypeTagReader
+import at.forsyte.apalache.io.lir.TypeTagReader
 
 /**
  * A json decoder, using the UJson representation
