+++ tla-io/src/main/scala/at/forsyte/apalache/io/lir/TlaType1PrinterPredefs.scala
@@ -1,0 +1,5 @@
+package at.forsyte.apalache.io.lir
+
+object TlaType1PrinterPredefs {
+  implicit val printer: TypeTagPrinter = new TlaType1TagPrinter
+}
