+++ tla-io/src/test/scala/at/forsyte/apalache/io/lir/TestCounterexampleWriter.scala
@@ -1,12 +1,13 @@
-package at.forsyte.apalache.tla.lir.io
+package at.forsyte.apalache.io.lir
 
-import java.io.{PrintWriter, StringWriter}
+import at.forsyte.apalache.tla.lir.UntypedPredefs._
 import at.forsyte.apalache.tla.lir._
 import at.forsyte.apalache.tla.lir.convenience.tla._
-import at.forsyte.apalache.tla.lir.UntypedPredefs._
 import org.junit.runner.RunWith
-import org.scalatest.junit.JUnitRunner
 import org.scalatest.FunSuite
+import org.scalatest.junit.JUnitRunner
+
+import java.io.{PrintWriter, StringWriter}
 
 @RunWith(classOf[JUnitRunner])
 class TestCounterexampleWriter extends FunSuite {
