--- tlair/src/main/scala/at/forsyte/apalache/tla/lir/io/TypeTagReader.scala
@@ -1,15 +1,0 @@
-package at.forsyte.apalache.tla.lir.io
-
-import at.forsyte.apalache.tla.lir.{TypeTag, Typed, Untyped}
-
-/**
- * Recovers TypeTags from their string representations.
- * Inverse of TypeTagPrinter.
- */
-trait TypeTagReader {
-  def apply(tagStr: String): TypeTag
-}
-
-object UntypedReader extends TypeTagReader {
-  override def apply(tagStr: String): TypeTag = Untyped()
-}
